qa/cephfs: fix TestRenameCommand and unmount the clinet before failing fs
We should else bring and wait for MDS to be up since it is needed
for unmounting of CephFS in CephFSTestCase.tearDown() to be successful,
or just unmount the mountpoints before failing the filesystem.

Since the mountpoint won't be used in later tests so we just unmount
it.

Fixes: https://tracker.ceph.com/issues/66946
Signed-off-by: Xiubo Li <xiubli@redhat.com>
(cherry picked from commit ae3faac004e64fa2e3bd4a3ff0d823db4da2685c)