qa/upgrade: use staggered upgrade features for reef-x/stress-split
This test was trying to partially upgrade the mons and OSDs by
kicking off an upgrade and then checking every 2 seconds if
enough had been upgraded. Since staggered upgrade parameters
were present in the initial reef release (not true for quincy)
it makes sense to use them instead in order to do this in a
more controlled manner.

Signed-off-by: Adam King <adking@redhat.com>