qa/suites/rados/thrash-old-clients/0-distros$: test on ubuntu_20.04 and drop nautilus
Centos 8 has gone end of life, so we need to choose a different distro on which
to test thrash-old-clients.

thrash-old-clients tests should only support N-3 releases. Nautilus fits with
this, but unfortunately there is no overlapping distro between nautilus, pacific,
octopus, AND quincy (bionic was dropped from quincy, and nautilus does not build
focal). As such, we are only able to test N-2.

Proof that focal is not available for octopus (this is where the test would search for packages):
https://shaman.ceph.com/api/search/?status=ready&project=ceph&flavor=default&distros=ubuntu%2F20.04%2Fx86_64&ref=nautilus

Signed-off-by: Laura Flores <lflores@ibm.com>