test/rgw: address potential race condition in reshard testing
The testing code does a `reshard add` followed by a `reshard list`. On
the narrow chance that the reshard was completed between those two
commands, we need to be able to handle the case where the reshard list
is empty. In that case we'll do a `bucket stats` to verify the reshard
completed successfully.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit aeaeaae7f5b677951773a6d718f327a9a04d5d75)