qa/crontab: update frequency and priority for rados nightlies
Although main rados nightlies are scheduled to run weekly, they
are so low in the queue that they take weeks to run, and when
they finally do, most of the jobs die (likely due to waiting in the
queue so long).

This commit increases the priority of main rados runs since
these are most important for debugging new failures. To compliment this
increase in priority, this commit also decreases the frequency of squid
rados runs from twice a week to once a week, since once is more than
enough. It also decreases the prioirity of squid runs.

Signed-off-by: Laura Flores <lflores@ibm.com>