msg, mon: introduce NONCE_WILDCARD to entity_addr_t
As `entity_addr_t` needs to be aware about the wildcard concept,
`Messenger::get_random_nonce()` delegates out the randomization
to it. However, there is a kludge: the space of numbers becomes
far smaller (32 bits instead of 64). However, effectively this
was always the case as `entity_addr_t::nonce` is typed as `__u32`.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>