run-make: install libstdc++-12-dev where gcc-12 is installed
debian builds don't depend on a specific version of gcc. ubuntu jammy's
gcc defaults to gcc-11, but gcc-12 packages are also available

we use clang for 'make check' on pull requests, and clang will use the
standard library from the most recent gcc version it finds. if gcc-12 is
installed but libstdc++-12-dev is not, the builds fail to link with:

> /usr/bin/ld: cannot find -lstdc++: No such file or directory

we don't want to make debian/control depend on a specific compiler
version, so i chose to add this to run-make.sh for run-make-check.sh
instead

Fixes: https://tracker.ceph.com/issues/67233

Signed-off-by: Casey Bodley <cbodley@redhat.com>