Single AsyncOpTracker per dispatcher
Removes the per layer trackers and replaces with a single tracker for image dispatch and another for object dispatch
Change register_dispatch interface so that it waits for the overall tracker to be idle before adding a new dispatch layer

Adding or removing a layer with IO in flight triggers a refresh operation to register or shutdown the changed layers.  The inflight IO is
queued waiting for the refresh to complete, but since layers cannot be added until the tracker is idle, IOs waiting for a refresh must be
removed from the tracker.  They are re-added to the tracker once IO is restarted after the refresh operation has completed.