mgr/dashboard: rbd table actions enhancements
Fixes:https://tracker.ceph.com/issues/67198

Below are the changes:-
Delete option moved to last in list.
Delete is disabled in case image is secondary.
Move to trash should be second last option.
Tooltip is added for ‘Move to Trash’ and ‘Delete’
Tooltip is added for Copy option
Subheading is added for Copy Form.
Tooltip is added for ‘Flatten’ option.
Delete tooltip content has been in case image is primary.

Signed-off-by: Prachi Goel <prachi@li-f0e8f2cc-27e0-11b2-a85c-b2772164128d.ibm.com>