Fixes:https://tracker.ceph.com/issues/67198 Signed-off-by: Prachi prgoel@redhat.com
Below are the changes:-
Delete option moved to last in list.
Delete is disabled in case image is not primary.
Move to trash should be second last option.
Tooltip is added for ‘Move to Trash’ and ‘Delete’
Tooltip is added for Copy option
Subheading is added for Copy Form.
Tooltip is added for ‘Flatten’ option.