rgw/h3: Listener dispatches packets in batches
each time the Listener calls recvmmsg() to read an array of packets, it
groups those packets by their associated Connection. this allows it to
deliver all of its packets to the Connection in a single call to
async_handle_packets()

this avoids the need for multiple round trips between the Listener's
executor and the Connection's. it also allows the Connection to process
all of the packets before calling writer_wake() to send response packets
or acks

Signed-off-by: Casey Bodley <cbodley@redhat.com>