mgr/smb: simplify orch backend enablement
We have a developer/debug module option that allows one to disable
triggering orchestration. When I tried to use it I thought it was
buggy and I had trouble diagnosing it. The mistake was on my side,
but the code change makes it much clearer what is being enabled
so I want to keep it.

Signed-off-by: John Mulligan <jmulligan@redhat.com>