rgw/d4n: implementation of list objects for a bucket (ordered listing) and list-object-versions.
Changes include the addition of an ordered set for
each bucket that stores the object, addition of etag
and object size to CacheObj structure needed for listing
objects, and addition of a test case to see the usage of
pipelining in Redis.  Objects are fetched from the cache
and the backend store and are then merged, before returning
the final list.

Signed-off-by: sinashan <sinatak1373@live.com>
Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>