mgr/vol: separate reusable helper code for reporting progress
A good amount of code from class CloneProgressReporter is generic to
getting, processing and then reporting stats through command output and
progress bar. This generic helper code can be used not just for stats of
async clone jobs but any async jobs. Therefore, move this code to a new
class in a new file so that it can be used, through class inheritance,
in both the classes, CloneProgressReporter as well as
PurgeProgressReporter (which will be written in following commits).

Signed-off-by: Rishabh Dave <ridave@redhat.com>