mds: do not trim segments after open file table commit
Previously, the trimming of expired segments would only occur if the open file
table's committed sequence number is past the segment to be trimmed. This is
now part of the expiry checks so it's no longer necessary to restart trimming
when the open file table commit completes.

Furthermore, this was confusing to `flush journal` as it was waiting for the
journal head write but the open file table commit already triggered the trim.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>