crimson/common/gated: Comment out ceph_assert for shard ID check
After adding sid (shard ID) to Gated, each shard should have its own gate.
However, some users of Gated had a single instance, which either failed
these assertions or created the Gated object on the wrong shard.
A tracker issue #67773 was opened for this, and this commit comments out
the assertions until the issue is resolved.

Signed-off-by: NitzanMordhai <nmordech@redhat.com>