cephadm: emit warning if daemon's image is not to be used
If an image is not specified, cephadm shell will use the image
corresponding to a Ceph daemon running on the host (and will log a
debug message to that effect).

However, it will only use that image if it appears in the output of:

podman images --filter label=ceph=True --filter dangling=false

This commit means that cephadm will emit a warning if the container
image it was going to use fails this check, so the operator has more
of a clue to what is going on.

Fixes: https://tracker.ceph.com/issues/67778
Signed-off-by: Matthew Vernon <mvernon@wikimedia.org>