ceph-crash: fix auth name usage and startup log noise
At startup ceph-crash ignored the predefined authentication names and
--name parameter at startup connect test, it only used the default admin
authentication.
This was not fatal because crash upload tried to find a working name and
only moved the crash file to the posted dir when the upload was
successful (potentially generating log spam in the process).

This commit changes this behaviour so ceph-crash will first try to find
a working authentication name from the predefined list or command line
switch (without log spam) and then use the first working one to upload
crash files (so this commit basically restores documented behaviour).

Fixes: https://tracker.ceph.com/issues/64102
Signed-off-by: Balazs Kalmar <kalmar15@gmail.com>