mgr/cephadm: open ceph-exporter when firewalld is enabled
ceph-exporter port was not opened properly during daemon deployment
forcing the user to open it manually. This change fixes the logic to
open the port automatically.

Fixes: https://tracker.ceph.com/issues/67975

Signed-off-by: Redouane Kachach <rkachach@ibm.com>
(cherry picked from commit fd93ecd88a5f31172ec99921f72bee12bdd4a6b9)

Conflicts:
	src/pybind/mgr/cephadm/tests/test_services.py

Conflict is because there was a new security related test that
was part of a feature in main but not squid that included the
ceph-exporter port