mgr/smb: accept public_addrs on cli when creating cluster
We can set the public ip address to set for the cluster using the
declarative method by providing the information in the resource
description. The corresponding functionality is not available with the
imperative method of creating the smb cluster.

This patch adds this functionality by allowing the user the option of
providing the a public address on the command line when creating the
smb cluster.

Signed-off-by: Sachin Prabhu <sp@spui.uk>