osd: erasure_code: Remove pointer from call to get_min_avail_to_read_shards
Following a review comment, observing that read_request is never null and
as such can be passed by reference, rather than by pointer.

A further review request suggested that this function should return
read_request in a tl::expected wrapper. However, this ends up with passing
this function more parameters to enable it to construct the read request,
thus making the code less efficient, so I did not implement this.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>