osd/PrimaryLogPG: Reuse clones when rolling back [Part 2/4]
In order to calculate new_last_overlap, we bring forward
each overlap between the rollback target to the newest clone we have.
We should begin with the full rollback target range and not from it's
clone_overlaps.

As the rollback target will replace the current head,
we should adjust it's clone_overlaps accordingly.
Mutating these entries is safe as the snapset ref is to new_snapset,
which is discarded on error.

Fixes [2/4]: https://tracker.ceph.com/issues/64735

Co-authored-by: Samuel Just <sjust@redhat.com>
Co-authored-by: dian Xing <xingd@chinatelecom.cn>
Signed-off-by: Matan Breizman <mbreizma@redhat.com>