osd/PrimaryLogPG: Reuse clones when rolling back [Part 1/4]
Skip the last clone_overlap when calculating intersections:
Rollback overwrites the head, so we don't care about the overlap between
the most recent clone and th head.
We want the overlap between the rollback target and the most recent clone.

Calculate the overlap based on:
<rollback target> - <most recent clone>
and not based on:
<rollback target> - <head>

Fixes [1/4]: https://tracker.ceph.com/issues/64735

Co-authored-by: Samuel Just <sjust@redhat.com>
Co-authored-by: dian Xing <xingd@chinatelecom.cn>
Signed-off-by: Matan Breizman <mbreizma@redhat.com>