osd/PrimaryLogPG: Reuse clones when rolling back [Part 3/3]
modified_ranges is used later on in PG::make_writeable to be
substracted from the newest_overlap (in the most recent clone) we have.
The modified_ranges resulted from a rollbacl operation should be only
the set of bytes that are not `new_last_overlap`.

---

head:   CCCC
clone2: AABB []
clone1: BBBB [2~2]

For example, rolling back to clone2 will result in
an *empty* modified range.
While rolling back to clone1 will reuslt in a partial
modified range of [0~2].

---

make_writeable is responsible for calculating the stats num_bytes field.
The above is achieved by the following:

1) modified_ranges.intersection_of(newest_overlap)
2) ctx->delta_stats.num_bytes += ctx->modified_ranges.size();

However, with rollbacks, the intersection of modified_ranges with the
newest_overlap will always be empty since we subtract one from another
in _do_rollback_to already.

Thus we need to accommodate num_bytes calculation to be
incremented in _do_rollback_to as well.

Fixes [3/3]: https://tracker.ceph.com/issues/64735

Signed-off-by: Matan Breizman <mbreizma@redhat.com>